refactor: extract GitHubService from GitService
